Tried to change as much to BigDecimal as possible. Also worked on adding more indicators.

+++ src/main/java/com/jarias14/tekstratego/service/pricer/biz/PricerService.java
@@ -1,5 +1,6 @@
 package com.jarias14.tekstratego.service.pricer.biz;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.SortedMap;
 
@@ -9,7 +10,7 @@ public interface PricerService {
     
     public Indicator createIndicator(Indicator indicator);
     public Indicator retrieveIndicator(String indicatorId);
-    public SortedMap<Calendar, Double> calculateIndicator(String indicatorId, String stockId, String startDate, String endDate);
-    public SortedMap<Calendar, Double> calculateIndicatorUsingNumberOfBars(String indicatorId, String stockId, String startDate, int numberOfBars);
+    public SortedMap<Calendar, BigDecimal> calculateIndicator(String indicatorId, String stockId, String startDate, String endDate);
+    public SortedMap<Calendar, BigDecimal> calculateIndicatorUsingNumberOfBars(String indicatorId, String stockId, String startDate, int numberOfBars);
 
 }
