+++ src/main/java/com/jarias14/tekstratego/service/thinker/dao/ThinkerDAO.java
@@ -1,5 +1,6 @@
 package com.jarias14.tekstratego.service.thinker.dao;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.List;
 import java.util.SortedMap;
@@ -14,9 +15,9 @@ public interface ThinkerDAO {
 
     public Hypothesis readHypothesis(String hypothesisId);
     
-    public SortedMap<Calendar, Double> getIndicatorValues(String indicatorId, Stock stock, Calendar startDate, Calendar endDate);
+    public SortedMap<Calendar, BigDecimal> getIndicatorValues(String indicatorId, Stock stock, Calendar startDate, Calendar endDate);
     
-    public SortedMap<Calendar, Double> getIndicatorValues(String indicatorId, Stock stock, Calendar today, Integer numberOfBars);
+    public SortedMap<Calendar, BigDecimal> getIndicatorValues(String indicatorId, Stock stock, Calendar today, Integer numberOfBars);
     
     public void postTradeAlers(String portfolioId, SortedMap<Calendar,List<TradeAlert>> tradeAlerts);
